reasoning/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
dspy-ai==2.4.9
openai==1.3.7
httpx==0.25.0
